requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
email-validator==2.1.0.post1
reportlab==4.1.0
openpyxl==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9 